Extract action logic into hook
